hw13/pyproject.toml
[tool.poetry]
name = "hw11"
version = "0.1.0"
description = ""
authors = ["PollyK21 <pol.kulish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.2.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.40.0"
sphinx = "^7.3.7"
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
